rename wrapper-attachment -> wrapper-meme-server, and initialize io-utils module
