+++ settings.gradle
@@ -11,12 +11,13 @@ include ':sphinx:application:common:logger'
 include ':sphinx:application:common:menus:menu-bottom'
 include ':sphinx:application:common:menus:menu-bottom-profile-pic'
 
-include ':sphinx:application:common:wrappers:wrapper-attachment'
 include ':sphinx:application:common:wrappers:wrapper-chat'
 include ':sphinx:application:common:wrappers:wrapper-common'
 include ':sphinx:application:common:wrappers:wrapper-contact'
 include ':sphinx:application:common:wrappers:wrapper-invite'
+include ':sphinx:application:common:wrappers:wrapper-io-utils'
 include ':sphinx:application:common:wrappers:wrapper-lightning'
+include ':sphinx:application:common:wrappers:wrapper-meme-server'
 include ':sphinx:application:common:wrappers:wrapper-message'
 include ':sphinx:application:common:wrappers:wrapper-message-media'
 include ':sphinx:application:common:wrappers:wrapper-relay'
