+++ sphinx/application/common/wrappers/wrapper-io-utils/src/main/java/chat/sphinx/wrapper_io_utils/InputStreamProvider.kt
@@ -1,0 +1,7 @@
+package chat.sphinx.wrapper_io_utils
+
+import java.io.InputStream
+
+abstract class InputStreamProvider {
+    abstract fun newInputStream(): InputStream
+}
