+++ sphinx/application/common/wrappers/wrapper-meme-server/build.gradle
