+++ sphinx/application/common/wrappers/wrapper-meme-server/src/main/java/chat/sphinx/wrapper_meme_server/AuthenticationToken.kt
@@ -1,4 +1,4 @@
-package chat.sphinx.wrapper_attachment
+package chat.sphinx.wrapper_meme_server
 
 @Suppress("NOTHING_TO_INLINE")
 inline fun String.toAuthenticationToken(): AuthenticationToken? =
