+++ sphinx/application/data/concepts/concept-meme-server/build.gradle
@@ -4,6 +4,6 @@ plugins {
 }
 
 dependencies {
-    api project(path: ':sphinx:application:common:wrappers:wrapper-attachment')
+    api project(path: ':sphinx:application:common:wrappers:wrapper-meme-server')
     api project(path: ':sphinx:application:common:wrappers:wrapper-message-media')
 }
