+++ sphinx/application/data/concepts/concept-meme-server/src/main/java/chat/sphinx/concept_meme_server/MemeServerTokenHandler.kt
@@ -1,6 +1,6 @@
 package chat.sphinx.concept_meme_server
 
-import chat.sphinx.wrapper_attachment.AuthenticationToken
+import chat.sphinx.wrapper_meme_server.AuthenticationToken
 import chat.sphinx.wrapper_message_media.token.MediaHost
 
 abstract class MemeServerTokenHandler {
