+++ sphinx/application/data/concepts/repositories/concept-repository-contact/build.gradle
@@ -7,6 +7,7 @@ dependencies {
     api project(path: ':sphinx:application:common:kotlin-response')
     api project(path: ':sphinx:application:common:wrappers:wrapper-contact')
     api project(path: ':sphinx:application:common:wrappers:wrapper-invite')
+    api project(path: ':sphinx:application:common:wrappers:wrapper-io-utils')
     implementation project(path: ':sphinx:application:common:wrappers:wrapper-message-media')
 
     implementation deps.kotlin.coroutinesCore
